fixed float values

+++ database/migrations/2019_09_28_125036_create_invoice_items_table.php
@@ -19,7 +19,7 @@ class CreateInvoiceItemsTable extends Migration
             $table->string('title');
             $table->integer('quantity');
             $table->string('unit');
-            $table->integer('unitCost');
+            $table->float('unitCost');
             $table->integer('invoice_id');
             $table->timestamps();
         });
